requirements.txt
bandit==1.7.0
boto3==1.17.20
botocore==1.20.20
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
decompyle3 @ git+https://github.com/rocky/python-decompile3@47b5587b0a18b9fa7bea901a3694b56cb9c7f073
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-SSE==1.0.0
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
greenlet==1.0.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
pbr==5.5.1
pefile==2019.4.18
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
PyYAML==5.4.1
redis==3.5.3
rq==1.7.0
s3transfer==0.3.4
six==1.15.0
smmap==3.0.5
spark-parser==1.8.9
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.8
stdlib-list==0.8.0
stevedore==3.3.0
uncompyle6==3.7.4
unpy2exe==0.4
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
xdis==5.0.7
yara-python==4.0.5
